+++ src/react/hooks/useLazyQuery.ts
@@ -1,6 +1,6 @@
 import { DocumentNode } from 'graphql';
 import { TypedDocumentNode } from '@graphql-typed-document-node/core';
-import { useCallback, useState } from 'react';
+import { useCallback, useEffect, useState } from 'react';
 
 import {
   LazyQueryHookOptions,
@@ -26,21 +26,36 @@ export function useLazyQuery<TData = any, TVariables = OperationVariables>(
   options?: LazyQueryHookOptions<TData, TVariables>
 ): QueryTuple<TData, TVariables> {
   const [execution, setExecution] = useState<
-    { called: boolean, options?: QueryLazyOptions<TVariables> }
+    {
+      called: boolean,
+      options?: QueryLazyOptions<TVariables>,
+      resolves: Array<(result: LazyQueryResult<TData, TVariables>) => void>,
+    }
   >({
     called: false,
+    resolves: [],
   });
 
   const execute = useCallback<
     QueryTuple<TData, TVariables>[0]
   >((executeOptions?: QueryLazyOptions<TVariables>) => {
+    let resolve!: (result: LazyQueryResult<TData, TVariables>) => void;
+    const promise = new Promise<LazyQueryResult<TData, TVariables>>(
+      (resolve1) => (resolve = resolve1),
+    );
     setExecution((execution) => {
       if (execution.called) {
         result && result.refetch(executeOptions?.variables);
       }
 
-      return { called: true, options: executeOptions };
+      return {
+        called: true,
+        resolves: [...execution.resolves, resolve],
+        options: executeOptions,
+      };
     });
+
+    return promise;
   }, []);
 
   let result = useQuery<TData, TVariables>(query, {
@@ -51,6 +66,13 @@ export function useLazyQuery<TData = any, TVariables = OperationVariables>(
     fetchPolicy: execution.called ? options?.fetchPolicy : 'standby',
     skip: undefined,
   });
+  useEffect(() => {
+    const { resolves } = execution;
+    if (!result.loading && resolves.length) {
+      setExecution((execution) => ({ ...execution, resolves: [] }));
+      resolves.forEach((resolve) => resolve(result));
+    }
+  }, [result, execution]);
 
   if (!execution.called) {
     result = {
@@ -66,7 +88,7 @@ export function useLazyQuery<TData = any, TVariables = OperationVariables>(
     for (const key of EAGER_METHODS) {
       const method = result[key];
       result[key] = (...args: any) => {
-        setExecution({ called: true });
+        setExecution((execution) => ({ ...execution, called: true }));
         return (method as any)(...args);
       };
     }